Changed systray window restore behavior
Previous I used show(); to show the MainWindow again but this does not "active" it. With active I mean to bring the window to the front and restore if from being minimized. This behavior is changed now and a click on the Systray Icon will make it visible instead of just appear in the taskbar (when it was minimized before).